requirements.txt
torch
scikit-learn
tqdm
numpy
timm
pandas
pillow
matplotlib
